\name{gbod-package}
\alias{gbod-package}
\alias{gbod}
\docType{package}
\title{
\packageTitle{gbod}
}
\description{
\packageDescription{gbod}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gbod}
\packageIndices{gbod}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gbod}

Maintainer: \packageMaintainer{gbod}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
